Rename correccion-ft references to generar-ft-de-fs
