+++ generar-ft-de-fs-application/src/main/java/com/comerzzia/bricodepot/generarftdefs/runner/GenerarFtDeFsRunner.java
@@ -1,6 +1,6 @@
 package com.comerzzia.bricodepot.generarftdefs.runner;
 
-import com.comerzzia.bricodepot.generarftdefs.services.correccion.CorreccionFtService;
+import com.comerzzia.bricodepot.generarftdefs.services.generacion.GenerarFtDeFsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
@@ -9,10 +9,10 @@ import java.io.BufferedReader;
 import java.io.FileReader;
 
 @Component
-public class CorreccionFtRunner implements CommandLineRunner {
+public class GenerarFtDeFsRunner implements CommandLineRunner {
 
     @Autowired
-    private CorreccionFtService servicio;
+    private GenerarFtDeFsService servicio;
 
     @Override
     public void run(String... args) throws Exception {
